feat: إضافة القرية والهاتف في الفواتير، كشف حساب، دعم الكسور في وحدات العمل، وإصلاح خطأ الحد الأدنى
- إضافة حقل village إلى نموذج Customer
- إظهار رقم الهاتف والقرية في جميع الفواتير والسندات
- إنشاء صفحة كشف حساب شامل للعملاء (print/statement/[id])
- إنشاء API كشف الحساب (api/customers/[id]/statement)
- تغيير workUnits من Int إلى Decimal لدعم الكسور
- إصلاح عدم تطابق حساب الحد الأدنى بين الواجهة والخادم
- إضافة زر كشف حساب في صفحة المشتركين
- إضافة حقل القرية ووحدات العمل في نموذج المشتركين
- إضافة تقرير فحص شامل للنظام (AUDIT_REPORT.md)
- إضافة migration SQL للتغييرات على قاعدة البيانات

+++ src/app/api/billing/[cycleId]/entries/route.ts
@@ -11,7 +11,7 @@ const batchEntriesSchema = z.object({
       customerId: z.string().optional(),
       currentReading: z.number().min(0, 'القراءة الحالية يجب أن تكون أكبر من أو تساوي الصفر'),
       consumption: z.number().min(0).optional(),
-      workUnits: z.number().int().min(0).optional(),
+      workUnits: z.number().min(0).optional(),
       serviceFee: z.number().min(0).optional(),
       fine: z.number().min(0).optional(),
       exemption: z.number().min(0).optional(),
